icde 26 submission, before revision

+++ src/kdtree/KDTree.cc
@@ -667,7 +667,7 @@ void SpatialIndex::KDTree::KDTree::intersectsWithQuery(const IShape& query, IVis
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
 }
-void SpatialIndex::KDTree::KDTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, uint64_t key_low, uint64_t key_high)
+void SpatialIndex::KDTree::KDTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, double key_low, double key_high)
 {
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
