+++ src/mvrtree/MVRTree.cc
@@ -318,7 +318,7 @@ void SpatialIndex::MVRTree::MVRTree::intersectsWithQuery(const IShape& query, IV
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
 }
-void SpatialIndex::MVRTree::MVRTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, uint64_t key_low, uint64_t key_high)
+void SpatialIndex::MVRTree::MVRTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, double key_low, double key_high)
 {
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
