+++ src/rtree/RTree.cc
@@ -729,7 +729,7 @@ void SpatialIndex::RTree::RTree::intersectsWithQuery(const IShape& query, IVisit
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
 }
-void SpatialIndex::RTree::RTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, uint64_t key_low, uint64_t key_high)
+void SpatialIndex::RTree::RTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, double key_low, double key_high)
 {
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
