+++ src/tprtree/TPRTree.cc
@@ -337,7 +337,7 @@ void SpatialIndex::TPRTree::TPRTree::intersectsWithQuery(const IShape& query, IV
 	rangeQuery(IntersectionQuery, query, v);
 }
 
-void SpatialIndex::TPRTree::TPRTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, uint64_t key_low, uint64_t key_high)
+void SpatialIndex::TPRTree::TPRTree::intersectsWithQueryLearnedIndex(const IShape& query, IVisitor& v, double key_low, double key_high)
 {
 	if (query.getDimension() != m_dimension) throw Tools::IllegalArgumentException("intersectsWithQuery: Shape has the wrong number of dimensions.");
 	rangeQuery(IntersectionQuery, query, v);
